Mutiple bug fixes, performance improvements and minor feature introductions and enhancements:
Updated the load form with record count
Added search feature in PluginEditor on request from Manish
Made the code window under the Plugin tab read-only
Added PluginResult.IsSignatureUnique method to allows plugins to check if a issue is already reported before reporting it
Updated the Is404 method of the crawler to reduce memory usage
Fixed a bug in the Scan Queue storage, stores FullUrl now instead of Url
Made changes to ScanManager to avoid aborting the ScanManager thread
Added 'Clear Shell Result' option based on request from Amol and Prasanna
Fixed a bug in the Proxy logging section that didn't store the edited response body properly
Updated the ThreadStore code to enable clearing of unused items quicker to reduce memory footprint
ThreadStore.Get now throws an exception if the requested item is not present in the store. Was returning a null value earlier.